requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.7
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
autotrain-advanced==0.8.2
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.8.2
boto3==1.34.141
botocore==1.34.141
Brotli==1.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==10.5.3
clarifai-grpc==10.6.1
click==8.1.7
codecarbon==2.3.5
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cpm-kernels==1.0.11
crewai==0.36.0
crewai-tools==0.4.8
cryptography==42.0.5
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.27.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==6.1.9
easyocr==1.7.1
einops==0.7.0
email_validator==2.2.0
embedchain==0.1.116
emoji==2.12.1
evaluate==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.5
ffmpy==0.3.2
filelock==3.15.1
filetype==1.2.0
FinGPT==0.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gpt4all==2.7.0
gptcache==0.1.43
gradio==4.37.2
gradio_client==1.0.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
hf_transfer==0.1.6
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflate64==1.0.0
iniconfig==2.0.0
inquirerpy==0.3.4
instructor==1.3.3
invisible-watermark==0.2.0
ipadic==1.0.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jiwer==3.0.3
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.7
langchain-cohere==0.1.8
langchain-community==0.2.7
langchain-core==0.2.12
langchain-huggingface==0.0.3
langchain-openai==0.1.14
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.77
latex2mathml==3.77.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.3.post0
loguru==0.7.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdtex2html==1.3.0
mdurl==0.1.2
memzero==0.0.7
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
ollama==0.2.1
onnxruntime==1.18.1
openai==1.34.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.82
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pfzy==0.3.4
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
posthog==3.5.0
pretty-errors==1.2.25
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
py7zr==0.21.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pylance==0.9.18
pyngrok==7.1.6
pynvml==11.5.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pypng==0.20220715.0
pyppmd==1.1.0
pyproject_hooks==1.1.0
pyreqwest_impersonate==0.4.9
pyright==1.1.370
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-rapidjson==1.18
pytube==15.0.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
pyzstd==0.16.0
qrcode==7.4.2
rapidfuzz==3.9.4
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
retry==0.9.2
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.1
rsa==4.9
ruff==0.5.0
s3transfer==0.10.2
sacremoses==0.1.1
safetensors==0.4.3
schema==0.7.5
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.10.1
seaborn==0.13.2
sec-api==1.0.18
selenium==4.22.0
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==2.5.1
seqeval==1.2.2
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
sympy==1.13.0rc1
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
termcolor==2.4.0
texttable==1.7.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.0.dev20240613
torchaudio==2.4.0.dev20240613
torchmetrics==1.4.0
torchvision==0.19.0.dev20240613
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.2
trio==0.26.0
trio-websocket==0.11.1
tritonclient==2.47.0
trl==0.9.4
tushare==1.4.7
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
ujson==5.10.0
Unidecode==1.3.8
unsloth @ git+https://github.com/unslothai/unsloth.git@36488b58677a2283de55aed0461df0237bb88fb8
unstructured==0.14.9
unstructured-client==0.24.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.17.4
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
